Modify bump-translations.bash script
- Rename download-translations.bash to bump-translations.bash
- Add git-commit command to the script
- Skip README.md file in empty translations check

+++ tools/i18n/bump-translations.bash
@@ -10,9 +10,18 @@ crowdin download -c crowdin.yml
 
 cd ../../src/_locales
 
+# Delete empty translations
 for lang in *; do
+  if [[ "$lang" == "README.md" ]]; then
+    continue
+  fi
   if [ "x$emptyTranslationFile" = "x$(cat $lang/messages.json)" ]; then
     echo "$lang has an empty translation. Deleting folder..."
     rm -rf $lang
+  else
+    git add $lang/messages.json
   fi
 done
+
+# Commit new translations
+git commit -m "Updating i18n files from Crowdin"
